HAL: fix clipboard abstraction and remove dupe in WindowsPlatformDialog
which also fixed the copy button on assertion dialogs (it was not working well after the first copy and had impact on system clipboard :p)

+++ Source/Runtime/Core/Public/HAL/Windows/WindowsPlatformMaths.h
@@ -258,7 +258,8 @@ public:
 #   ifdef ARCH_X64
         return ::__popcnt64(v);
 #   else
-        return (::__popcnt(u32(v)) + ::__popcnt(u32(v >> 32)));
+        return (u64(::__popcnt(u32(v))) +
+                u64(::__popcnt(u32(v >> 32))) );
 #   endif
     }
 };
